backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-Cors==4.0.1
gunicorn==21.2.0
requests==2.32.3
Werkzeug==3.0.3
psycopg[binary]==3.2.13
python-dotenv==1.0.1
PyJWT==2.9.0
reportlab
psycopg2-binary
celery[redis]==5.3.6
celery[redis]
sentry-sdk[flask]==2.8.0
opentelemetry-api==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-instrumentation-flask==0.47b0
opentelemetry-instrumentation-sqlalchemy==0.47b0
opentelemetry-exporter-otlp-proto-http==1.26.0
